requirements.txt
boto3 ==1.16.61
python-dateutil == 2.8.1
Jinja2 == 2.11.2
plotly == 4.10.0
botocore~=1.18.1